kream-auth-server/requirements.txt
Flask==2.3.3
Flask-Cors==4.0.0
bcrypt==4.0.1
PyJWT==2.8.0
